requirements.txt
gensim==4.3.2
nltk==3.8.1
numpy==1.26.3
matplotlib==3.8.2
pandas==2.1.4
pickle5==0.0.11
python-dotenv==1.0.0
scikit-learn==1.3.2
torch==2.1.2
seaborn==0.13.1